Change from yarn to npm for CI builds (#14)
This PR adds a small change on the CI pipeline to use the same tool on
the CI as we are using on local development.

`npm` suits better the needs of the project (where some inner inspecting
of the `node_modules` directory stored deps might be needed for custom
components inspecting). We are trying to offset `yarn`'s better
performance with some specific parameters and options.