Add basic navigation structure (#26)
This PR includes some starting foundations for the site:

- It adds some basic styling such as colours, headings, adjust the
element's global radius… And it also includes the [`Inter`
typography](https://rsms.me/inter/) in a variable set, to be used along
all the site. It also updates the site to show Kotlin code snippets
properly highlighted (colours theme to be updated in the future).

- It makes the necessary changes in the navigation bar and the footer to
include the proper links outlined in the site wireframes. It does so
while also updating the pages placeholder template to make these pages
use navbar and footer now.

- It configures and start using a method to use YAML (`*.yml` files) as
a data source for pages. I think this go in line with the front matter
data for Markdown files in the docs and blog pages, which might make it
easier for non-technical people to make changes in case they deemed
necessary. The code for this is also very clean, and it looks exactly
the same way as we import styles for the pages, which makes it
reasonable.