Add Incubation section (#46)
This PR adds the initial structure and config changes to allow a new
parallel section to the _Learn_ documents called _Incubation_.

The structure to its own sidebar will be autogenerated in the same way
as the one in the _Learn_ section, by creating the different documents
and directories tree structure. Also, the section name and initial
document is fully open to be modified to your likings. If you do modify
it, just try to correct the different manual links (navbar and footer,
basically), so the build is able to complete.

The position and method to show this new category is temporarily being
shown as a new item in the same level as the Learn section. I leave to
your criteria and @israelpzglez's proposal the final way to include this
section in the navbar and footer links.


![image](https://user-images.githubusercontent.com/7753447/222213878-15565333-d1c4-42df-9283-d4df68a48d38.png)


![image](https://user-images.githubusercontent.com/7753447/222214357-aff8906c-0dbb-4dd8-b067-edb857fe7e8c.png)