GitHub Pages setup (#115)
This pull request creates a new GitHub Actions workflow to deploy the
Arrow website in GitHub Pages. The workflow will be triggered when a new
release is published. Anyway, we can't enable GitHub Pages yet because
the feature is not available for private repositories.

Additionally, it also upgrades the Node version to the 18 major version.