Change Λrrow occurrences for Arrow (#146)
This PR modifies all occurrences of referring to the library name as
`Λrrow` for its proper name: `Arrow`.

While the usage of the uppercase lambda is cool, and it creates a nice
effect on the text, it totally defeats the purpose of having accessible
web content. And it also brings some inconsistencies and problems when
writing the library on different media, etc. It also creates the issue
of the library name not being properly indexed by web crawlers and bots.

While some possible bypass for those problems could be the use of the
[`aria-label`
property](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-label),
it would really be a hacky shortcut misusing it, as [that wouldn't be
the goal of that HTML
property](https://auroratide.com/posts/how-to-not-use-aria-label). And
all this while not really solving the underlying issue of having an
improper (in terms of pure text) library name.

The usage of the uppercased lambda name, `Λrrow,` should only be used on
display media like images, videos, and such, but never text content.

This closes #145