Improve page layouts for responsive (#152)
### Description:

This PR makes some adjustments to all views on site pages to improve
their layouts on smaller devices like tablets and mobile. There's still
some pending work to do, as the build step of the site is messing some
styles (I need to check a bit on the webpack plugins minimizing stuff
and such to find the culprit). It also changes a bit the way the
`StarsCount` component was being shown to use an SVG icon instead of the
star ⭐ emoji.

It also removes some unused code that was there commented or directly
unused.

This fixes #150. 


### Preview:


![imagen](https://user-images.githubusercontent.com/7753447/229835275-29917a97-1ec8-4c59-84fa-d674fec9d2c3.png)


![imagen](https://user-images.githubusercontent.com/7753447/229835541-8b7d6f5e-42f0-4ce0-8f76-46bc5d07fa62.png)