Adjust custom CSS specificity affected by Infima (#158)
### Description

This PR make some changes on some selectors' specificity. Due to the way
Docusaurus build production styles, this could sometimes create a
problem through the CSS injection order, making some Infima styles
override custom ones. To overcome this, we are using a subtle trick by
duplicating the class in our CSS, to increase the specificity of our
selectors.

It doesn't seem like an easy task to solve in general for Docusaurus.
More info in: https://github.com/facebook/docusaurus/issues/3678

It also adds proper styling for the Table of Contents on mobile.

This closes #147

### Preview


![imagen](https://user-images.githubusercontent.com/7753447/230122746-a8eb2d3d-f70b-4bb1-92b2-1ea06ca65563.png)

![imagen](https://user-images.githubusercontent.com/7753447/230122431-ddd3dd06-5b06-4e8b-a91e-1aae1723e5f2.png)